refact

+++ sources/Panel/src/Hardware/HAL/HAL_DAC2_stm.cpp
@@ -44,7 +44,7 @@ void HAL_DAC2::Init()
 }
 
 
-void HAL_DAC2::SetValue(uint value)
+void HAL_DAC2::SetValue(int value)
 {
-    HAL_DAC_SetValue(&handle, DAC1_CHANNEL_2, DAC_ALIGN_8B_R, value);
+    HAL_DAC_SetValue(&handle, DAC1_CHANNEL_2, DAC_ALIGN_8B_R, static_cast<uint>(value));
 }
